requirements.txt
allure-behave==2.9.43
allure-python-commons==2.9.43
Appium-Python-Client==1.3.0
attrs==21.2.0
bcrypt==3.2.0
behave==1.2.6
boto3==1.20.24
botocore==1.23.24
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
colorama==0.4.4
comtypes==1.1.10
configparser==5.2.0
crayons==0.4.0
cryptography==36.0.0
idna==3.3
jmespath==0.10.0
jsonpath==0.82
mysqlclient==2.1.0
numpy==1.21.4
paramiko==2.8.1
parse==1.19.0
parse-type==0.5.2
pluggy==1.0.0
pycparser==2.21
pymongo==3.12.0
PyNaCl==1.4.0
python-dateutil==2.8.2
pytz==2021.3
PyYAML==5.4.1
requests==2.26.0
s3transfer==0.5.0
selenium==4.8.2
six==1.16.0
sshtunnel==0.4.0
urllib3==1.26.7
webdriver-manager==3.5.2
undetected_chromedriver==3.4.6